+++ packages/core/config/webpack.config.js
@@ -27,13 +27,7 @@ module.exports = (env, argv) => {
         rules: client.getRules(),
       },
       plugins: client.getPlugins(),
-      optimization: {
-        splitChunks: {
-          name: 'common',
-          chunks: 'all',
-        },
-        runtimeChunk: isProd,
-      },
+      optimization: client.getOptimization(),
     },
     {
       context: buildContext,
@@ -72,11 +66,7 @@ module.exports = (env, argv) => {
         rules: server.getRules(),
       },
       plugins: server.getPlugins(),
-      optimization: {
-        // This keeps the emitted code readable if we need to review it manually.
-        // Minimization isn't useful for NodeJS anyways.
-        minimize: false,
-      },
+      optimization: server.getOptimization(),
     },
   ];