+++ packages/core/config/webpack/plugins.js
@@ -1,4 +1,5 @@
 const webpack = require('webpack');
+const MiniCSSExtractPlugin = require('mini-css-extract-plugin');
 const CleanPlugin = require('clean-webpack-plugin');
 const { StatsWriterPlugin } = require('webpack-stats-plugin');
 const { maybeResolveUserModule } = require('../../utils/userModule');
@@ -38,6 +39,10 @@ module.exports = function getPlugins(context) {
           maxChunks: 1,
         }),
         new webpack.HashedModuleIdsPlugin(),
+        new MiniCSSExtractPlugin({
+          filename: '[name].[hash].css',
+          chunkFilename: '[id].[hash].css',
+        }),
       ];
 
     case 'development_server':
@@ -46,6 +51,10 @@ module.exports = function getPlugins(context) {
         // Ensures async components can be rendered sync server-side.
         new webpack.optimize.LimitChunkCountPlugin({
           maxChunks: 1,
+        }),
+        new MiniCSSExtractPlugin({
+          filename: '[name].css',
+          chunkFilename: '[id].css',
         }),
       ];
 
@@ -68,6 +77,10 @@ module.exports = function getPlugins(context) {
           noSources: true,
           publicPath: `${rootUrl}/`,
         }),
+        new MiniCSSExtractPlugin({
+          filename: '[name].[hash].css',
+          chunkFilename: '[id].[hash].css',
+        }),
       ];
 
     case 'development_client':
@@ -75,6 +88,10 @@ module.exports = function getPlugins(context) {
         ...commonPlugins,
         new webpack.NamedModulesPlugin(),
         new webpack.HotModuleReplacementPlugin(),
+        new MiniCSSExtractPlugin({
+          filename: '[name].css',
+          chunkFilename: '[id].css',
+        }),
       ];
 
     default: