+++ packages/styled-components/components/html/themes/unstyled.js
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+/* eslint-disable import/prefer-default-export */
+export const HTMLWrapper = styled.div``;
+/* eslint-enable */