+++ packages/postcss/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.14.0](https://github.com/alleyinteractive/irving/packages/integrations/compare/v2.14.0-beta.1...v2.14.0) (2020-06-15)
+
+**Note:** Version bump only for package @irvingjs/postcss
+
+
+
+
+
+# [2.14.0-beta.1](https://github.com/alleyinteractive/irving/packages/integrations/compare/v2.14.0-beta.0...v2.14.0-beta.1) (2020-06-15)
+
+**Note:** Version bump only for package @irvingjs/postcss
+
+
+
+
+
+# [2.13.0](https://github.com/alleyinteractive/irving/packages/integrations/compare/v2.13.0-rc.0...v2.13.0) (2020-06-12)
+
+**Note:** Version bump only for package @irvingjs/postcss
+
+
+
+
+
+# [2.13.0-rc.0](https://github.com/alleyinteractive/irving/packages/integrations/compare/v2.12.1...v2.13.0-rc.0) (2020-06-12)
+
+**Note:** Version bump only for package @irvingjs/postcss
+
+
+
+
+
+## [2.12.1](https://github.com/alleyinteractive/irving/packages/integrations/compare/v2.12.0...v2.12.1) (2020-06-10)
+
+**Note:** Version bump only for package @irvingjs/postcss
+
+
+
+
+
+# [2.12.0](https://github.com/alleyinteractive/irving/packages/integrations/compare/v2.11.1...v2.12.0) (2020-06-10)
+
+**Note:** Version bump only for package @irvingjs/postcss
+
+
+
+
+
 ## [2.11.1](https://github.com/alleyinteractive/irving/packages/integrations/compare/v2.11.0...v2.11.1) (2020-06-03)
 
 **Note:** Version bump only for package @irvingjs/postcss